feat: relocate auth
