+++ src/main/java/com/clipboarder/clipboarder/controller/ClipboarderUserController.java
@@ -1,13 +1,7 @@
 package com.clipboarder.clipboarder.controller;
 
-import com.clipboarder.clipboarder.entity.dto.ClipboarderUserDTO;
-import com.clipboarder.clipboarder.entity.dto.SignupResponse;
 import com.clipboarder.clipboarder.service.ClipboarderUserService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,12 +10,5 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/user")
 public class ClipboarderUserController {
     private final ClipboarderUserService clipboarderUserService;
-
-    @PostMapping
-    public ResponseEntity<SignupResponse> signUp(@RequestBody @Valid ClipboarderUserDTO clipboarderUserDTO){
-        clipboarderUserService.signUp(clipboarderUserDTO);
-
-        return ResponseEntity.ok().body(new SignupResponse(true, "GOOD~~"));
-    }
 
 }
