+++ src/main/java/com/clipboarder/clipboarder/controller/TextController.java
@@ -1,14 +1,16 @@
 package com.clipboarder.clipboarder.controller;
 
-import com.clipboarder.clipboarder.entity.dto.SignupResponse;
-import com.clipboarder.clipboarder.entity.dto.TextCopyResponse;
+import com.clipboarder.clipboarder.entity.dto.response.SignUpResponseDTO;
+import com.clipboarder.clipboarder.entity.dto.response.TextCopyResponse;
 import com.clipboarder.clipboarder.entity.dto.TextDTO;
-import com.clipboarder.clipboarder.entity.dto.TextGetResponse;
+import com.clipboarder.clipboarder.entity.dto.response.TextGetResponse;
 import com.clipboarder.clipboarder.service.TextService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -37,12 +39,12 @@ public class TextController {
         return ResponseEntity.ok(new TextGetResponse(true, textDTOList));
     }
 
-    @DeleteMapping
-    public ResponseEntity<SignupResponse> delete(HttpServletRequest request, @RequestBody TextDTO textDTO){
-        String token = request.getHeader("Authorization");
-
-        textService.delete(token, textDTO);
-
-        return ResponseEntity.ok(new SignupResponse(true, "GOOD~"));
-    }
+//    @DeleteMapping
+//    public ResponseEntity<SignUpResponseDTO> delete(HttpServletRequest request, @RequestBody TextDTO textDTO){
+//        String token = request.getHeader("Authorization");
+//
+//        textService.delete(token, textDTO);
+//
+//        return ResponseEntity.ok(new SignUpResponseDTO(true, "GOOD~"));
+//    }
 }
